1. move bean to MultipleNamespacesAzureServiceBusMessagingAutoConfiguration.java 2. implements CommandLineRunner to auto send request
