+++ servicebus/spring-cloud-azure-starter-integration-servicebus/multiple-namespaces/src/main/java/com/azure/spring/sample/servicebus/QueueSendController.java
@@ -24,23 +24,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class QueueSendController {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(QueueSendController.class);
-    private static final String OUTPUT_CHANNEL = "queue1.output";
+    private static final String OUTPUT_CHANNEL_QUEUE1 = "queue1.output";
     private static final String QUEUE_NAME = "queue1";
 
-    @Autowired
-    QueueSendGateway messagingGateway;
-
-    /**
-     * Posts a message to a Service Bus Queue
-     */
-    @PostMapping("/queues")
-    public String send(@RequestParam("message") String message) {
-        this.messagingGateway.send(message);
-        return message;
-    }
-
     @Bean
-    @ServiceActivator(inputChannel = OUTPUT_CHANNEL)
+    @ServiceActivator(inputChannel = OUTPUT_CHANNEL_QUEUE1)
     public MessageHandler queueMessageSender(ServiceBusTemplate serviceBusTemplate) {
         serviceBusTemplate.setDefaultEntityType(ServiceBusEntityType.QUEUE);
         DefaultMessageHandler handler = new DefaultMessageHandler(QUEUE_NAME, serviceBusTemplate);
@@ -61,9 +49,9 @@ public class QueueSendController {
 
     /**
      * Message gateway binding with {@link MessageHandler}
-     * via {@link MessageChannel} has name {@value OUTPUT_CHANNEL}
+     * via {@link MessageChannel} has name {@value OUTPUT_CHANNEL_QUEUE1}
      */
-    @MessagingGateway(defaultRequestChannel = OUTPUT_CHANNEL)
+    @MessagingGateway(defaultRequestChannel = OUTPUT_CHANNEL_QUEUE1)
     public interface QueueSendGateway {
         void send(String text);
     }
